erased GameRoom page
